Get Exif data

Returns Exchangeable image file format data for a single photograph. For more information on how Exif data differs from metadata see: https://www.flickr.com/services/api/flickr.photos.getExif.html

get_exif(photo_id = NULL)

Arguments

photo_id
Id of photograph

Value

A dataframe of "exchangeable image file format" information for the given photograph

Details

When running the function you need an API key saved as photosearcher_key.sysdata in your working directory. If this is the first function you run you will be prompted to create and enter your API key. The API key will then be saved as photosearcher_key.sysdata in your working directory and is used for all function.

See also

Examples

# NOT RUN {
get_exif(photo_id = 48704764812)
# }