Get related tags

Takes a tag and returns the top tags related to that tag.

related_tags(tag)

Arguments

tag
character. tag to search.

Value

character. Tags most associated with input tag.

Details

Uses the flickr.tags.getRelated API method from the Flickr API. See https://www.flickr.com/services/api/flickr.tags.getRelated.html for more information on the API method.

When running the function you need an API key saved as photosearcher_key.sysdata in your working directory. If this is the first function you run you will be prompted to create and enter your API key. The API key will then be saved as photosearcher_key.sysdata in your working directory and is used for all function.

Examples

if (FALSE) {
related_tags(tag = "car")

related_tags(tag = "monkey")

related_tags(tag = "river")
}